refactor(cli): split execute.go by concern (E3)
Move the broker-denial classification, agent_directive protocol, and
response printers out of execute.go into execute_broker.go,
execute_directive.go, and execute_print.go. execute.go keeps the command,
request build/send, and operation resolution. Intra-package moves only.
