+++ cli/internal/cli/api/execute.go
@@ -8,7 +8,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"net"
 	"net/http"
 	"net/url"
 	"os"
@@ -21,7 +20,6 @@ import (
 	"github.com/jentic/jentic-one/cli/internal/cli/clictx"
 	"github.com/jentic/jentic-one/cli/internal/cli/ux"
 	"github.com/jentic/jentic-one/cli/internal/config"
-	"github.com/jentic/jentic-one/cli/internal/theme"
 	"github.com/spf13/cobra"
 )
 
@@ -359,32 +357,6 @@ func (a *app) executeE(cmd *cobra.Command, opts *executeOptions, target string) 
 	return a.executeOutput(cmd, opts, resp)
 }
 
-// brokerTLSMismatch reports whether err is the "https client hit an http server"
-// signature against a loopback broker — the local default-scheme papercut UX-4
-// makes actionable. Restricted to loopback so we never suggest downgrading a
-// remote broker to plaintext.
-func brokerTLSMismatch(err error, brokerURL string) bool {
-	if err == nil || !strings.Contains(err.Error(), "server gave HTTP response to HTTPS client") {
-		return false
-	}
-	u, perr := url.Parse(brokerURL)
-	if perr != nil {
-		return false
-	}
-	return isLoopbackHostname(u.Hostname())
-}
-
-// isLoopbackHostname reports whether host is "localhost" or a loopback IP.
-func isLoopbackHostname(host string) bool {
-	if host == "localhost" {
-		return true
-	}
-	if ip := net.ParseIP(host); ip != nil {
-		return ip.IsLoopback()
-	}
-	return false
-}
-
 // badFlagKV builds the coded error for a malformed key=value flag (ARCH-4). A
 // bad --path/--query/--header value is agent-causable INPUT, so it must surface
 // a machine error_code (MISSING_ARGUMENT) with an actionable hint — not a bare
@@ -517,274 +489,4 @@ func (a *app) resolveOperation(cmd *cobra.Command, opts *executeOptions, target 
 		return nil, errInspectMissingFields
 	}
 	return &opInfo, nil
-}
-
-func (a *app) executeOutput(cmd *cobra.Command, opts *executeOptions, resp *http.Response) error {
-	if opts.raw {
-		// Read (bounded by the broker transport's cap) then redact before
-		// streaming, so --raw matches the redaction guarantee of the JSON and
-		// `jentic api` paths (SEC-2). A secret in an upstream body must not leak
-		// just because the caller asked for raw output.
-		body, err := io.ReadAll(resp.Body)
-		if err != nil {
-			return fmt.Errorf("read response: %w", err)
-		}
-		if _, err := a.Out.Write(ux.RedactBytes(body)); err != nil {
-			return err
-		}
-		if isBrokerDenial(resp) {
-			return brokerDeniedErr(resp)
-		}
-		return nil
-	}
-
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return fmt.Errorf("read response: %w", err)
-	}
-
-	if jsonOrPretty(cmd, opts.json) {
-		if err := a.executeJSONOutput(resp, respBody); err != nil {
-			return err
-		}
-	} else {
-		a.executePrettyOutput(resp, respBody)
-	}
-
-	// A broker denial (403/409/424/401) means the call did not run; exit 2 so a
-	// scripted agent can branch on the denial instead of mistaking the 4xx body
-	// for success. The exit code keys off the *status*, not the presence of an
-	// agent_directive: some denials (e.g. action_denied from a permission rule)
-	// carry no directive, and gating exit 2 on a parsed directive would let those
-	// silently exit 0 — the exact regression this surfacing is meant to remove.
-	// When a directive *is* present it enriches the message with recovery steps.
-	if isBrokerDenial(resp) {
-		if directive, ok := parseAgentDirective(resp, respBody); ok {
-			a.printAgentDirective(directive)
-		} else {
-			// No agent_directive on the body (UX7): a first-timer who most needs the
-			// pointer would otherwise get only the generic "broker denied" line and
-			// a dead end. Synthesize a default next-step from the HTTP status the
-			// broker already returned so no denial is a dead end.
-			a.printSynthesizedDenialRecovery(resp.StatusCode)
-		}
-		return brokerDeniedErr(resp)
-	}
-	return nil
-}
-
-// printSynthesizedDenialRecovery prints a best-effort recovery hint for a broker
-// denial that carried no agent_directive, keyed off the HTTP status the broker
-// already returned (UX7). It mirrors printAgentDirective's shape (instruction +
-// `run:` command) so a directive-less denial reads the same as a directed one.
-// Unknown statuses fall back to the generic setup check.
-func (a *app) printSynthesizedDenialRecovery(status int) {
-	fmt.Fprintln(a.Err, theme.Warn.Render("Denied — recovery required:"))
-	switch status {
-	case http.StatusForbidden: // 403: have an identity, no access to this toolkit yet.
-		fmt.Fprintln(a.Err, "  This agent isn't bound to the toolkit you called. Check what you can run, then request access.")
-		fmt.Fprintln(a.Err, "  run: "+theme.Accent.Render("jentic access whoami"))
-		fmt.Fprintln(a.Err, "  run: "+theme.Accent.Render("jentic access request --toolkit <vendor/name> --wait"))
-	case http.StatusFailedDependency: // 424: no credential provisioned for the call.
-		fmt.Fprintln(a.Err, "  No credential is provisioned for this call. Provision one, then retry.")
-		fmt.Fprintln(a.Err, "  run: "+theme.Accent.Render("jentic access request --toolkit <vendor/name> --provision --wait"))
-	case http.StatusUnauthorized: // 401: credential expired / needs reconnecting.
-		fmt.Fprintln(a.Err, "  Your credential needs reconnecting. Re-run access to refresh it.")
-		fmt.Fprintln(a.Err, "  run: "+theme.Accent.Render("jentic access request --toolkit <vendor/name> --provision --wait"))
-	default:
-		fmt.Fprintln(a.Err, "  The broker denied this call. Check what you can run and your setup.")
-		fmt.Fprintln(a.Err, "  run: "+theme.Accent.Render("jentic access whoami"))
-	}
-	// Point at the read-only self-check as the catch-all when the specific hint
-	// above doesn't unblock (UX9).
-	fmt.Fprintln(a.Err, "  stuck? "+theme.Accent.Render("jentic doctor"))
-}
-
-// brokerDeniedErr is the typed denial every broker-denial exit shares (AGT-6):
-// BROKER_DENIED (exit 2) with the denying HTTP status in Details, so the agent
-// envelope and the exit code come from the same table. The response body /
-// directive recovery text is printed by the caller before this is returned.
-func brokerDeniedErr(resp *http.Response) *ux.CodedError {
-	return &ux.CodedError{
-		Code:    ux.CodeBrokerDenied,
-		Msg:     "the broker denied this call before it reached the upstream API",
-		Details: map[string]any{"http_status": resp.StatusCode},
-	}
-}
-
-// isBrokerDenial reports whether a response is one the broker itself emitted to
-// deny a call the agent can recover from: missing toolkit binding → 403,
-// ambiguous toolkit → 409, credential needs reconnect → 401, no credential →
-// 424. Each carries an agent_directive (see broker/web/errors.STATUS_BY_ERROR).
-//
-// Status alone is NOT sufficient: the broker is a transparent forward proxy, so
-// an *upstream* API can return these same 4xx codes on a call the broker
-// successfully proxied (the upstream auth failed, the resource is forbidden,
-// etc.). Treating those as broker denials would exit 2 and print a misleading
-// "recovery required" for a call that actually ran. The broker disambiguates
-// with the Jentic-Error-Origin response header (broker/core/headers): it stamps
-// "broker" on its own errors and "upstream" on mirrored pass-through 4xx/5xx
-// (broker/services/execution/pipeline.enrich_error_origin). So a denial-class
-// status is a broker denial only when the origin is not "upstream" (a missing
-// header is treated as broker, since the loopback broker always sets it on its
-// own errors and only a non-conformant proxy would omit it).
-func isBrokerDenial(resp *http.Response) bool {
-	if resp == nil {
-		return false
-	}
-	if errorOrigin(resp) == errorOriginUpstream {
-		return false
-	}
-	switch resp.StatusCode {
-	case http.StatusUnauthorized,
-		http.StatusForbidden,
-		http.StatusConflict,
-		http.StatusFailedDependency:
-		return true
-	default:
-		return false
-	}
-}
-
-// errorOriginUpstream is the Jentic-Error-Origin value the broker stamps on a
-// mirrored upstream response (broker ErrorOrigin.UPSTREAM). The matching header
-// name mirrors broker/core/headers.JenticHeader.ERROR_ORIGIN.
-const errorOriginUpstream = "upstream"
-
-func errorOrigin(resp *http.Response) string {
-	if resp == nil {
-		return ""
-	}
-	return strings.ToLower(strings.TrimSpace(resp.Header.Get("Jentic-Error-Origin")))
-}
-
-// agentDirective mirrors the broker's problem+json agent_directive extension
-// member (broker/core/exceptions.py AgentDirective). The strategy vocabulary
-// below mirrors broker AgentStrategy; until the contract is a shared OpenAPI
-// schema, the Python test test_directive_factories_emit_known_strategies and
-// this list must be kept in lock-step (review P1-1).
-type agentDirective struct {
-	Strategy    string         `json:"strategy"`
-	Parameters  map[string]any `json:"parameters"`
-	Instruction string         `json:"human_readable_instruction"`
-}
-
-// Recovery strategies the broker may emit in an agent_directive (mirrors
-// broker AgentStrategy): wait, retry, modify_headers, prompt_human,
-// switch_toolkit, fatal. Only the ones this CLI branches on are named here.
-const (
-	directiveWait  = "wait"
-	directiveRetry = "retry"
-)
-
-// parseAgentDirective extracts an agent_directive from a denial response body.
-// It only treats recoverable broker-denial responses as directives so a normal
-// 4xx (including an upstream pass-through with an incidental
-// "agent_directive"-shaped body) can't trip the exit code.
-func parseAgentDirective(resp *http.Response, body []byte) (agentDirective, bool) {
-	if !isBrokerDenial(resp) {
-		return agentDirective{}, false
-	}
-	var envelope struct {
-		Directive *agentDirective `json:"agent_directive"`
-	}
-	if err := json.Unmarshal(body, &envelope); err != nil || envelope.Directive == nil {
-		return agentDirective{}, false
-	}
-	return *envelope.Directive, true
-}
-
-// printAgentDirective renders a recovery directive to stderr, lifting the
-// suggested_command / provisioning_url out of parameters so the agent (or its
-// operator) sees the exact next action without parsing JSON.
-func (a *app) printAgentDirective(d agentDirective) {
-	fmt.Fprintln(a.Err, theme.Warn.Render("Denied — recovery required:"))
-	if d.Instruction != "" {
-		fmt.Fprintln(a.Err, "  "+d.Instruction)
-	}
-	if cmd, ok := d.Parameters["suggested_command"].(string); ok && cmd != "" {
-		fmt.Fprintln(a.Err, "  run: "+theme.Accent.Render(cmd))
-	}
-	if u, ok := d.Parameters["provisioning_url"].(string); ok && u != "" {
-		fmt.Fprintln(a.Err, "  open: "+theme.Accent.Render(u))
-	}
-	if cands, ok := d.Parameters["candidates"].([]any); ok && len(cands) > 0 {
-		parts := make([]string, 0, len(cands))
-		for _, c := range cands {
-			if s, isStr := c.(string); isStr {
-				parts = append(parts, s)
-			}
-		}
-		if len(parts) > 0 {
-			fmt.Fprintln(a.Err, "  candidates: "+theme.Accent.Render(strings.Join(parts, ", ")))
-		}
-	}
-	// A wait/retry directive carries a backoff hint the agent should honor before
-	// retrying; surface it so the recovery loop doesn't hot-spin.
-	if d.Strategy == directiveWait || d.Strategy == directiveRetry {
-		if secs, ok := d.Parameters["retry_after_seconds"]; ok {
-			fmt.Fprintf(a.Err, "  retry after: %v\n", secs)
-		}
-	}
-	// Catch-all self-check for when the directive's step doesn't unblock (UX9).
-	fmt.Fprintln(a.Err, "  stuck? "+theme.Accent.Render("jentic doctor"))
-}
-
-func (a *app) executeJSONOutput(resp *http.Response, body []byte) error {
-	headers := make(map[string]string)
-	for k := range resp.Header {
-		headers[k] = resp.Header.Get(k)
-	}
-
-	var parsedBody any
-	if err := json.Unmarshal(body, &parsedBody); err != nil {
-		parsedBody = string(body)
-	}
-
-	envelope := map[string]any{
-		// AGT-23: stamp the machine-contract schema_version like every sanctioned
-		// wrapper (ux.List/Page/Result/Export), so an agent can branch on the
-		// envelope shape. execute builds an ad-hoc map (not a ux type) because its
-		// body is the arbitrary upstream response, so the version is set here.
-		"schema_version": apiEnvelopeSchemaVersion,
-		"status":         resp.StatusCode,
-		"headers":        headers,
-		"body":           parsedBody,
-	}
-	if execID := resp.Header.Get("Jentic-Execution-Id"); execID != "" {
-		envelope["execution_id"] = execID
-	}
-
-	return writeJSON(a.Out, envelope)
-}
-
-func (a *app) executePrettyOutput(resp *http.Response, body []byte) {
-	statusLine := fmt.Sprintf("HTTP %d %s", resp.StatusCode, http.StatusText(resp.StatusCode))
-	switch {
-	case resp.StatusCode >= 200 && resp.StatusCode < 300:
-		fmt.Fprintln(a.Out, theme.Success.Render(statusLine))
-	case resp.StatusCode >= 400:
-		fmt.Fprintln(a.Out, theme.Warn.Render(statusLine))
-	default:
-		fmt.Fprintln(a.Out, statusLine)
-	}
-
-	for k, vs := range resp.Header {
-		if strings.HasPrefix(k, "Jentic-") {
-			fmt.Fprintln(a.Out, theme.Dim.Render(fmt.Sprintf("  %s: %s", k, strings.Join(vs, ", "))))
-		}
-	}
-
-	fmt.Fprintln(a.Out)
-	if len(body) > 0 {
-		// Redact the upstream body before display (SEC-2): the pretty path is
-		// human-facing but can still carry secrets echoed by an upstream API.
-		var pretty bytes.Buffer
-		if err := json.Indent(&pretty, body, "", "  "); err == nil {
-			_, _ = a.Out.Write(ux.RedactBytes(pretty.Bytes()))
-		} else {
-			_, _ = a.Out.Write(ux.RedactBytes(body))
-		}
-		fmt.Fprintln(a.Out)
-	}
 }
